ReadMe file added

+++ backend/E-master/src/main/java/com/example/E_master/services/CourseService.java
@@ -16,7 +16,13 @@ public class CourseService {
     @Autowired
     private CourseRepository courseRepository;
 
-    public Course createCourse(Course course, MultipartFile file) throws IOException {
+//    public Course createCourse(Course course, MultipartFile file) throws IOException {
+//        String filePath = "path/to/upload/directory/" + file.getOriginalFilename();
+//        file.transferTo(new File(filePath));
+//        course.setPdfPath(filePath);
+//        return courseRepository.save(course);
+//    }
+   public Course createCourse(Course course, MultipartFile file) throws IOException {
         String filePath = "path/to/upload/directory/" + file.getOriginalFilename();
         file.transferTo(new File(filePath));
         course.setPdfPath(filePath);
@@ -26,7 +32,7 @@ public class CourseService {
     public List<Course> getAllCourses() {
         return courseRepository.findAll();
     }
-    public Optional<Course> findById(Integer id) {
+    public Optional<Course> findById(Long id) {
         return courseRepository.findById(id);
     }
 }
